list and delete staff

+++ app/Http/Controllers/CityController.php
@@ -14,10 +14,8 @@ class CityController extends Controller
         $cities= City::with('cityManager')->get();
        
         if (request()->ajax()) {
-            
-            // $query = City::with('cityManager')->select('staff.*');
             return datatables()->of($cities)
-        ->addColumn('cityManagers', function (City $city) {
+               ->addColumn('cityManagers', function (City $city) {
             return $city->cityManager->name;
         })
                ->addColumn('action', function ($data) {
@@ -35,7 +33,7 @@ class CityController extends Controller
     ]
         );
     }
-
+//--------------------------- add new city -----------------------------
     public function store(Request $request)
     {
         $cityId = $request->id;
@@ -53,7 +51,7 @@ class CityController extends Controller
                          
         return Response()->json($city);
     }
-
+//--------------------- Edit name of city ------------------------------
     public function edit(Request $request)
     {
         $where = array('id' => $request->id);
@@ -61,12 +59,21 @@ class CityController extends Controller
       
         return Response()->json($city);
     }
-
+//-----------------------  delete city -------------------------------------
     public function destroy(Request $request)
     {
-        //  $gyms = Gym::Where("city_id","=",$request->id);
-        // if(empty($gyms))
-        $city = City::where('id', $request->id)->delete();
-        return Response()->json($city);
+        $gyms = City::find($request->id)->gyms;
+        $gymnumber =0 ;
+        foreach($gyms as $gym)
+        {
+            if(!empty($gym))
+            $gymnumber += 1;
+        }
+        if($gymnumber==0)
+        {
+            $city = City::where('id', $request->id)->delete();
+            return Response()->json($city);
+        }
+        
     }
 }
